RFCT: move image manager to image_process for maintainability
